Crate uu_tail

Re-exports

pub use args::uu_app;

Modules

args

chunks

Iterating over a file by chunks, either starting at the end of the file with ReverseChunks or at the end of piped stdin with LinesChunk or BytesChunk.

text

Functions

uumain