Struct uu_tail::chunks::LinesChunk

pub struct LinesChunk { /* private fields */ }

Works similar to a BytesChunk but also stores the number of lines encountered in the current buffer. The size of the buffer is limited to a fixed size number of bytes.

Implementations

impl LinesChunk

pub fn new(delimiter: u8) -> Self

pub fn has_data(&self) -> bool

Returns true if this buffer has stored any bytes.

Examples

let mut chunk = LinesChunk::new(b'\n');
assert!(!chunk.has_data());

chunk.buffer[0] = 1;
assert!(!chunk.has_data());

chunk.bytes = 1;
assert!(chunk.has_data());

pub fn get_buffer(&self) -> &[u8]

Returns this buffer safely. See BytesChunk::get_buffer

returns: &u8 with length self.bytes

pub fn get_buffer_with(&self, offset: usize) -> &[u8]

Returns this buffer safely with an offset applied. See BytesChunk::get_buffer_with.

returns: &u8 with length self.bytes - offset

pub fn get_lines(&self) -> usize

Return the number of lines the buffer contains. self.lines needs to be set before the call to this function returns the correct value. If the calculation of lines is needed then use self.count_lines.

pub fn fill(&mut self, filehandle: &mut impl BufRead) -> UResult<Option<usize>>

Fills self.buffer with maximal BUFFER_SIZE number of bytes, draining the reader by that number of bytes. This function works like the BytesChunk::fill function besides that this function also counts and stores the number of lines encountered while reading from the filehandle.

pub fn print_lines(&self, writer: &mut impl Write, offset: usize) -> UResult<()>

Print the bytes contained in this buffer calculated with the given offset in number of lines.

Arguments

writer: must implement Write

offset: An offset in number of lines.

pub fn print_bytes(&self, writer: &mut impl Write, offset: usize) -> UResult<()>

Print the bytes contained in this buffer beginning from the given offset in number of bytes.

Arguments

writer: must implement Write

offset: An offset in number of bytes.